% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{nudibranch_palette}
\alias{nudibranch_palette}
\title{Get a nudibranch colour palette}
\usage{
nudibranch_palette(name = NULL, n = NULL, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of the palette (e.g. "hypselodoris"). If NULL, prints
all available palette names.}

\item{n}{Number of colours to return. If NULL, returns all colours.}

\item{type}{Either "discrete" or "continuous".}
}
\value{
A character vector of hex colour codes.
}
\description{
Get a nudibranch colour palette
}
\examples{
nudibranch_palette("hypselodoris")
nudibranch_palette("elysia", n = 3)
}
