% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{nudibranch_palettes}
\alias{nudibranch_palettes}
\title{Nudibranch colour palettes}
\format{
An object of class \code{list} of length 12.
}
\usage{
nudibranch_palettes
}
\description{
A named list of colour palettes inspired by nudibranchs
and sea slugs photographed in Australian waters.
}
\keyword{datasets}
