% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{print.nudibranch_palette}
\alias{print.nudibranch_palette}
\title{Print a nudibranch palette}
\usage{
\method{print}{nudibranch_palette}(x, ...)
}
\arguments{
\item{x}{A nudibranch_palette object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print a nudibranch palette
}
