% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_nudibranch}
\alias{scale_colour_nudibranch}
\title{ggplot2 colour scale using a nudibranch palette}
\usage{
scale_colour_nudibranch(name, n = NULL, type = "discrete", ...)
}
\arguments{
\item{name}{Name of the palette (e.g. "hypselodoris")}

\item{n}{Number of colours. If NULL, returns all colours in the palette.}

\item{type}{Either "discrete" or "continuous"}

\item{...}{Additional arguments passed to ggplot2}
}
\description{
ggplot2 colour scale using a nudibranch palette
}
